Add '**' as an option for proxies, to trust any IP address (#46, #49)
However, if you are in the situation where, say, you have a
Content Distribution Network (like Amazon CloudFront) that passes to load
balancer (like Amazon ELB) then you may end up with a chain of unknown
proxies forwarding from one to another. In that case, '*' above would
only match the final proxy (the load balancer in this case) which means
that calling `$request->getClientIp()` would return the IP address
of the next proxy in line (in this case one of the Content Distribution
Network ips) rather than the original client IP.To always get the original
client IP, you need to trust all the proxies in the route to your request.

You can do this by setting proxies to ['0.0.0.0/0', '2000:0:0:0:0:0:0:0/3']
which means all IP addresses will be trusted.

This patch implements ** as syntactic sugar for ['0.0.0.0/0', '2000:0:0:0:0:0:0:0/3']