Fixes for no_trailing_whitespace_in_comment
I am absolutely being completely nit picky, but I am also trying to cut down on the number of files in my second commit in every project which is after running `php-cs-fixer`. As there is mixed usage in this file, I thought I would take the opportunity, as I am operating under the belief that the framework already checks for this.